fingerprint: extract shared value/function canonicalizer; strengthen digest
Move the value canonicalizer and deep function fingerprint (_canon_value,
_fingerprint_function, _fingerprint_codeobj, _referenced_global_names,
_safe_repr, _Unfingerprintable) out of compile_kernel into a shared
numbox/utils/fingerprint.py; compile_kernel._formula_fingerprint imports them.

Upgrade numbox.utils.digest to fingerprint plain functions with the shared
closure/global-aware walker (cloudpickle fallback for codeless callables and
un-canonicalizable closures), so two callbacks with identical source but
different captured closure-cell / referenced-global values key distinctly --
which the previous bare code-object hash missed. Add the fingerprint automodule
to the utils docs.
