+++ numbox/core/variable/compile_kernel.py
@@ -23,10 +23,8 @@ uncached (no anchor, no numba cache) -- never reused, never wrong.
 import hashlib
 import warnings
 
-from types import CodeType, FunctionType, ModuleType
+from types import FunctionType
 from typing import Any, Callable
-
-import numpy as np
 
 from numba import njit, typeof
 from numba.core.ccallback import CFunc
@@ -45,6 +43,9 @@ from numbox.core.variable.utils import (
 from numbox.core.variable.variable import (
     QUAL_SEP, CompiledGraph, CompiledNode, Graph, Variable, make_qual_name,
 )
+from numbox.utils.fingerprint import (
+    _Unfingerprintable, _canon_value, _fingerprint_function, _safe_repr,
+)
 from numbox.utils.preprocessing import (
     _anchor_root, _materialize_anchor, _orphan_anchor_sweep,
 )
@@ -60,94 +61,6 @@ def _effective_flags(jit_options: dict | None) -> dict:
     as `fastmath`/`error_model` otherwise diverge across the discovery boundary)."""
     opts = {**_default_jit_options, **(jit_options or {})}
     return {k: v for k, v in opts.items() if k != "cache"}
-
-
-class _Unfingerprintable(Exception):
-    """A value the cache digest cannot canonicalize; the kernel goes uncached."""
-
-
-def _safe_repr(obj: object) -> str:
-    """``repr(obj)`` that never raises -- the fingerprint fallback must always
-    yield a string so an un-fingerprintable formula degrades to uncached rather
-    than crashing when its ``__repr__`` itself raises."""
-    try:
-        return repr(obj)
-    except Exception:  # noqa: BLE001 - fallback must not crash on a raising __repr__
-        return f"<{type(obj).__name__} repr-failed>"
-
-
-def _canon_value(value: Any, seen: set[int]) -> str:
-    if value is None or isinstance(value, (bool, int, float, complex, str, bytes)):
-        return repr(value)
-    if isinstance(value, np.generic):
-        return f"npscalar({value.dtype.str};{value.tobytes().hex()})"
-    if isinstance(value, np.ndarray):
-        if value.dtype.hasobject:
-            raise _Unfingerprintable(f"object-dtype ndarray {value.dtype.str}")
-        data = np.ascontiguousarray(value)
-        try:
-            raw = hashlib.sha256(data.tobytes()).hexdigest()
-        except (ValueError, TypeError) as e:
-            raise _Unfingerprintable(f"unhashable ndarray {value.dtype.str}") from e
-        return f"ndarray({data.dtype.str};{value.shape};{raw})"
-    if isinstance(value, (tuple, list)):
-        return f"{type(value).__name__}[" + ",".join(_canon_value(v, seen) for v in value) + "]"
-    if isinstance(value, (set, frozenset)):
-        return f"{type(value).__name__}[" + ",".join(sorted(_canon_value(v, seen) for v in value)) + "]"
-    if isinstance(value, dict):
-        items = sorted((_canon_value(k, seen), _canon_value(v, seen)) for k, v in value.items())
-        return "dict[" + ",".join(f"{k}:{v}" for k, v in items) + "]"
-    if isinstance(value, ModuleType):
-        return f"module({value.__name__})"
-    if isinstance(value, Dispatcher):
-        topts = _canon_value(dict(getattr(value, "targetoptions", {}) or {}), seen)
-        return f"dispatcher({_fingerprint_function(value.py_func, seen)};{topts})"
-    if isinstance(value, FunctionType):
-        return f"function({_fingerprint_function(value, seen)})"
-    raise _Unfingerprintable(type(value).__name__)
-
-
-def _fingerprint_codeobj(code: CodeType, seen: set[int]) -> str:
-    consts = ",".join(
-        _fingerprint_codeobj(c, seen) if isinstance(c, CodeType) else _canon_value(c, seen)
-        for c in code.co_consts
-    )
-    return (
-        f"code({code.co_code.hex()};flags={code.co_flags};argc={code.co_argcount};"
-        f"kwonly={code.co_kwonlyargcount};names={','.join(code.co_names)};consts=[{consts}])"
-    )
-
-
-def _referenced_global_names(code: CodeType) -> set[str]:
-    names = set(code.co_names)
-    for c in code.co_consts:
-        if isinstance(c, CodeType):
-            names |= _referenced_global_names(c)
-    return names
-
-
-def _fingerprint_function(func: FunctionType, seen: set[int]) -> str:
-    if id(func) in seen:
-        return f"recursive({func.__qualname__})"
-    seen = seen | {id(func)}
-    code = func.__code__
-    cells = []
-    for name, cell in zip(code.co_freevars, func.__closure__ or ()):
-        try:
-            contents = cell.cell_contents
-        except ValueError as e:
-            raise _Unfingerprintable("empty closure cell") from e
-        cells.append(f"{name}={_canon_value(contents, seen)}")
-    hashed_globals = []
-    for name in sorted(_referenced_global_names(code)):
-        if name in func.__globals__:
-            hashed_globals.append(f"{name}={_canon_value(func.__globals__[name], seen)}")
-    return (
-        f"func({func.__module__}:{func.__qualname__};{_fingerprint_codeobj(code, seen)};"
-        f"defaults={_canon_value(func.__defaults__ or (), seen)};"
-        f"kwdefaults={_canon_value(func.__kwdefaults__ or {}, seen)};"
-        f"closure=[{';'.join(cells)}];globals=[{';'.join(hashed_globals)}])"
-    )
 
 
 def _formula_fingerprint(formula) -> tuple[str, bool]:
